requirements.txt
ipykernel
openpyxl
pandas
tabula-py
JPype1
pdfquery
mysql-connector-python